Separate better-sqlite3 and ioredis dependencies into dedicated packages
This change removes optional database dependencies from @atxp/common to reduce
bundle size and dependency overhead for users who don't need these features.

**New Packages:**
- @atxp/sqlite-db: SQLite database implementation with better-sqlite3
- @atxp/redis-db: Redis OAuth database implementation with ioredis

**Changes:**
- Remove better-sqlite3 and ioredis from @atxp/common dependencies
- Move SQLite platform abstractions to @atxp/sqlite-db
- Move Redis OAuth database code to @atxp/redis-db
- Update tests to use MemoryOAuthDb for in-memory testing
- Update factory functions to guide users to appropriate packages
- Remove SQLite-specific platform tests from common package

**Migration:**
- For SQLite: import from '@atxp/sqlite-db'
- For Redis: import from '@atxp/redis-db'
- Core functionality remains in '@atxp/common'
