+++ packages/atxp-client/src/oauth.expo.test.ts
@@ -1,8 +1,8 @@
 // Expo/React Native-specific tests for platform abstraction
-// Only include tests relevant to cross-platform or Expo/React Native logic
+// Only include tests relevant to cross-platform logic
 
 import { describe, it, expect } from 'vitest';
-import { getIsReactNative, crypto, sqlite } from '@atxp/common';
+import { getIsReactNative, crypto } from '@atxp/common';
 
 describe('Platform Abstraction (Expo)', () => {
   it('should detect React Native environment', () => {
@@ -17,12 +17,5 @@ describe('Platform Abstraction (Expo)', () => {
     const hash = await crypto.digest(testData);
     expect(hash).toBeInstanceOf(Uint8Array);
     expect(hash.length).toBeGreaterThan(0);
-  });
-
-  it('should have working SQLite implementation', () => {
-    const db = sqlite.openDatabase(':memory:');
-    expect(db).toBeDefined();
-    expect(typeof db.execAsync).toBe('function');
-    expect(typeof db.prepareAsync).toBe('function');
   });
 });
