+++ packages/atxp-common/src/index.ts
@@ -1,9 +1,7 @@
 export * from './jwt.js';
 export * from './logger.js'
-export * from './oAuthDb.js'
 export * from './memoryOAuthDb.js';
 export * from './oAuthDbFactory.js';
-export * from './redisOAuthDb.js';
 export * from './oAuthResource.js'
 export * from './paymentRequiredError.js'
 export * from './servers.js';
